Fix:Kotlinのキャストエラーの修正

+++ packages/nekonata_location_fetcher/android/src/main/kotlin/com/example/nekonata_location_fetcher/NekonataLocationFetcherPlugin.kt
@@ -85,8 +85,8 @@ class NekonataLocationFetcherPlugin: FlutterPlugin, MethodCallHandler {
   private fun configure(call: MethodCall) {
     call.argument<String>("notificationTitle")?.let { Store.notificationTitle = it }
     call.argument<String>("notificationText")?.let { Store.notificationText = it }
-    call.argument<Float>("distanceFilter")?.let { Store.distanceFilter = it }
-    call.argument<Long>("interval")?.let { Store.interval = it }
+    call.argument<Double>("distanceFilter")?.let { Store.distanceFilter = it.toFloat() }
+    call.argument<Int>("interval")?.let { Store.interval = it.toLong() }
   }
 
   private fun start() {
